requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pymongo==4.6.1
